subscriber/requirements.txt
aioredis==1.3.1
astroid==2.4.2
async-timeout==3.0.1
asyncpg==0.21.0
autopep8==1.5.4
click==7.1.2
fastapi==0.61.2
flake8==3.8.4
h11==0.11.0
hiredis==1.1.0
httptools==0.1.1
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
pycodestyle==2.6.0
pydantic==1.7.2
pyflakes==2.2.0
PyJWT==1.7.1
pylint==2.6.0
python-dotenv==0.15.0
PyYAML==5.3.1
six==1.15.0
starlette==0.13.6
toml==0.10.2
uvicorn==0.12.2
uvloop==0.14.0
watchgod==0.6
websockets==8.1
wrapt==1.12.1
